+++ lib/rubocop/cop/style/trivial_accessors.rb
@@ -43,27 +43,39 @@ module Rubocop
           TrivialAccessors.config['AllowPredicates']
         end
 
+        def whitelist
+          whitelist = TrivialAccessors.config['Whitelist']
+          Array(whitelist).map(&:to_sym) + [:initialize]
+        end
+
         def predicate?(method_name)
           method_name[-1] == '?'
         end
 
         def trivial_reader?(method_name, args, body)
-          return false unless args.children.size == 0
+          looks_like_trivial_reader?(args, body) &&
+            !allowed_method?(method_name, body)
+        end
 
-          return false unless body && body.type == :ivar
-
-          return false if allow_predicates? && predicate?(method_name)
-
-          exact_name_match? ? names_match?(method_name, body) : true
+        def looks_like_trivial_reader?(args, body)
+          args.children.size == 0 && body && body.type == :ivar
         end
 
         def trivial_writer?(method_name, args, body)
-          return false unless args.children.size == 1 &&
-            body && body.type == :ivasgn &&
-            body.children[1] && body.children[1].type == :lvar &&
-            method_name != :initialize
+          looks_like_trivial_writer?(args, body) &&
+            !allowed_method?(method_name, body)
+        end
 
-          exact_name_match? ? names_match?(method_name, body) : true
+        def looks_like_trivial_writer?(args, body)
+          args.children.size == 1 &&
+            body && body.type == :ivasgn &&
+            body.children[1] && body.children[1].type == :lvar
+        end
+
+        def allowed_method?(method_name, body)
+          allow_predicates? && predicate?(method_name) ||
+            whitelist.include?(method_name) ||
+            exact_name_match? && !names_match?(method_name, body)
         end
 
         def names_match?(method_name, body)